requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
custom-rocksdb @ file:///home/mengke/python-rocksdb/rocksdb_pybinding
exceptiongroup==1.3.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pyparsing==3.2.5
PyYAML==6.0.3
requests==2.32.5
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
torch==2.9.1
torch-geometric==2.7.0
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.6.2
xxhash==3.6.0
yarl==1.22.0
